Backend - Implement features in movements of money and control of registry in transactions 🚀

+++ backend/src/main/java/com/semillero/Finavex/dto/movementsMoney/RequestRegistrySaveMoney.java
@@ -3,8 +3,10 @@ package com.semillero.Finavex.dto.movementsMoney;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
+import java.math.BigDecimal;
+
 public record RequestRegistrySaveMoney(
-        Double savedAmount,
+        BigDecimal savedAmount,
         String note,
 
         @NotBlank(message = "El email no puede estar vacio")
