+++ backend/src/main/java/com/semillero/Finavex/services/movementsS/CurrencyFormatter.java
@@ -1,6 +1,8 @@
 package com.semillero.Finavex.services.movementsS;
 
 import org.springframework.stereotype.Component;
+
+import java.math.BigDecimal;
 import java.text.NumberFormat;
 import java.util.Locale;
 
@@ -8,10 +10,10 @@ import java.util.Locale;
 public class CurrencyFormatter {
 
     /**
-     * Formatea un Double a moneda sin símbolo ($)
+     * Formatea un BigDecimal a moneda sin símbolo ($)
      * Ejemplo: 14000.00 -> "14.000,00"
      */
-    public String formatCurrencyWithoutSymbol(Double amount) {
+    public String formatCurrencyWithoutSymbol(BigDecimal amount) {
         if (amount == null) {
             return "0,00";
         }
